move config module to separate folder
