+++ src/commands/inherit.js
@@ -1,5 +1,5 @@
 import { spawnSync } from 'child_process'
-import { getTask } from '../config.js'
+import { getTask } from '../config/config.js'
 import { logger } from '../logger/logger.js'
 
 export async function inherit() {
