+++ src/runTask.js
@@ -2,7 +2,7 @@ import { spawn } from 'cross-spawn'
 import kleur from 'kleur'
 import path, { relative } from 'path'
 import stripAnsi from 'strip-ansi'
-import { getDiffPath, getManifestPath, getNextManifestPath, getTask } from './config.js'
+import { getDiffPath, getManifestPath, getNextManifestPath, getTask } from './config/config.js'
 import { existsSync, mkdirSync, readFileSync, renameSync, unlinkSync, writeFileSync } from './fs.js'
 import { logger } from './logger/logger.js'
 import { computeManifest } from './manifest/computeManifest.js'
